+++ ShowerController/Model/TechnicalInformation.swift
@@ -15,16 +15,17 @@ class TechnicalInformation {
     @Relationship
     var device: Device?
 
-    private(set) var valveType: UInt8
-    private(set) var valveSoftwareVersion: UInt8
-    private(set) var uiType: UInt8
-    private(set) var uiSoftwareVersion: UInt8
-    private(set) var bluetoothType: UInt8
-    private(set) var bluetoothSoftwareVersion: UInt8
+    private(set) var valveType: UInt16
+    private(set) var valveSoftwareVersion: UInt16
+    private(set) var uiType: UInt16
+    private(set) var uiSoftwareVersion: UInt16
+    private(set) var bluetoothType: UInt16
+    private(set) var bluetoothSoftwareVersion: UInt16
 
-    init(valveType: UInt8, valveSoftwareVersion: UInt8, uiType: UInt8, uiSoftwareVersion: UInt8, bluetoothType: UInt8, bluetoothSoftwareVersion: UInt8) {
+    init(valveType: UInt16, valveSoftwareVersion: UInt16, uiType: UInt16, uiSoftwareVersion: UInt16, bluetoothType: UInt16, bluetoothSoftwareVersion: UInt16) {
         self.valveType = valveType
         self.valveSoftwareVersion = valveSoftwareVersion
+        // 42 = dual shower, 44 = shower+bath
         self.uiType = uiType
         self.uiSoftwareVersion = uiSoftwareVersion
         self.bluetoothType = bluetoothType
